requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
ldif3==3.1.1
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-card-me==0.9.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
surlex==0.2.0
toml==0.10.2
unicodecsv==0.14.1
urllib3==1.26.7
